+++ packages/api/src/commands/index.ts
@@ -3,6 +3,7 @@ export * from "./fs"
 export * from "./server"
 export * from "./tools"
 export * from "./extension"
+export * from "./system"
 export * from "./store"
 export * as db from "./db"
 export { JarvisExtDB } from "./db"
